requirements.txt
asgiref==3.11.1
Django==6.0.3
django-unfold==0.84.0
pillow==12.1.1
sqlparse==0.5.5
tzdata==2025.3
